third_party/WebKit/LayoutTests/svg/custom/keySplines-parsing-error-expected.txt

CONSOLE ERROR: Error: Invalid value for <animate> attribute keySplines=";;"
CONSOLE ERROR: Error: Invalid value for <animate> attribute keySplines="0 ,0  1 , 1  ;;"
Tests parsing of keySplines attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
